pyproject.toml
[project]
name = "quote-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "altair>=5",
    "google-genai>=1.67.0",
    "openpyxl>=3.1.5",
    "pandas>=2.0,<3",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.2",
    "streamlit>=1.40",
    "watchdog>=6.0.0",
]
